+++ src/frontend/apps/drive/src/features/explorer/utils/mimeTypes.ts
@@ -8,6 +8,7 @@ import mimePowerpoint from "@/assets/files/icons/mime-powerpoint.svg";
 import mimeAudio from "@/assets/files/icons/mime-audio.svg";
 import mimeVideo from "@/assets/files/icons/mime-video.svg";
 import mimeSQLite from "@/assets/files/icons/mime-sqlite.svg";
+import mimeGrist from "@/assets/files/icons/mime-grist.svg";
 
 import mimeCalcMini from "@/assets/files/icons/mime-calc-mini.svg";
 import mimeDocMini from "@/assets/files/icons/mime-doc-mini.svg";
@@ -20,6 +21,7 @@ import mimeVideoMini from "@/assets/files/icons/mime-video-mini.svg";
 import mimeArchiveMini from "@/assets/files/icons/mime-archive-mini.svg";
 import mimeSuspicious from "@/assets/files/icons/suspicious_file.svg";
 import mimeSQLiteMini from "@/assets/files/icons/mime-sqlite-mini.svg";
+import mimeGristMini from "@/assets/files/icons/mime-grist-mini.svg";
 
 import mimeArchive from "@/assets/files/icons/mime-archive.svg";
 import { getExtension } from "../utils/utils";
@@ -36,6 +38,7 @@ export enum MimeCategory {
   ARCHIVE = "archive",
   SUSPICIOUS = "suspicious",
   SQLITE = "sqlite",
+  GRIST = "grist",
 }
 
 export const MIME_TO_ICON = {
@@ -50,6 +53,7 @@ export const MIME_TO_ICON = {
   [MimeCategory.ARCHIVE]: mimeArchive,
   [MimeCategory.SUSPICIOUS]: mimeSuspicious,
   [MimeCategory.SQLITE]: mimeSQLite,
+  [MimeCategory.GRIST]: mimeGrist,
 };
 
 export const MIME_TO_ICON_MINI = {
@@ -64,6 +68,7 @@ export const MIME_TO_ICON_MINI = {
   [MimeCategory.ARCHIVE]: mimeArchiveMini,
   [MimeCategory.SUSPICIOUS]: mimeSuspicious,
   [MimeCategory.SQLITE]: mimeSQLiteMini,
+  [MimeCategory.GRIST]: mimeGristMini,
 };
 
 export const ICONS = {
@@ -79,6 +84,7 @@ export const ICONS = {
     [MimeCategory.ARCHIVE]: mimeArchiveMini,
     [MimeCategory.SUSPICIOUS]: mimeSuspicious,
     [MimeCategory.SQLITE]: mimeSQLiteMini,
+    [MimeCategory.GRIST]: mimeGristMini,
   },
   normal: {
     [MimeCategory.CALC]: mimeCalc,
@@ -92,6 +98,7 @@ export const ICONS = {
     [MimeCategory.ARCHIVE]: mimeArchive,
     [MimeCategory.SUSPICIOUS]: mimeSuspicious,
     [MimeCategory.SQLITE]: mimeSQLite,
+    [MimeCategory.GRIST]: mimeGrist,
   },
 };
 
@@ -106,6 +113,7 @@ export const MIME_TO_FORMAT_TRANSLATION_KEY = {
   [MimeCategory.VIDEO]: "mime.video",
   [MimeCategory.ARCHIVE]: "mime.archive",
   [MimeCategory.SQLITE]: "mime.sqlite",
+  [MimeCategory.GRIST]: "mime.grist",
 };
 
 export const MIME_MAP = {
@@ -156,6 +164,11 @@ export const getMimeCategory = (
     CALC_EXTENSIONS.includes(extension)
   ) {
     return MimeCategory.CALC;
+  }
+
+  // Special case: a SQLITE file that has a .grist extension is a grist file
+  if (mimetype === "application/vnd.sqlite3" && extension === "grist") {
+    return MimeCategory.GRIST;
   }
 
   if (MIME_TO_CATEGORY[mimetype]) {